Separate multimodal backend assembly
